[android] peripherals: move implementation to platform folder.
