+++ cmake/treedata/android/subdirs.txt
@@ -1,7 +1,6 @@
 xbmc/cores/RetroPlayer/process/android  cores/RetroPlayer/process/android
 xbmc/input/touch                        input/touch
 xbmc/input/touch/generic                input/touch/generic
-xbmc/peripherals/bus/android            peripherals/bus/android
 xbmc/powermanagement/android            powermanagement/android
 xbmc/filesystem/posix                   filesystem/posix
 xbmc/utils/posix                        utils/posix
@@ -13,4 +12,5 @@ xbmc/platform/linux/peripherals         platform/linux/peripherals
 xbmc/platform/android/activity          android/activity
 xbmc/platform/android/bionic_supplement android/bionicsupplement
 xbmc/platform/android/network           platform/android/network
+xbmc/platform/android/peripherals       platform/android/peripherals
 xbmc/platform/android/storage           platform/android/storage
