+++ xbmc/platform/android/peripherals/AndroidJoystickState.cpp
