+++ xbmc/platform/android/peripherals/AndroidJoystickState.h
