+++ xbmc/platform/android/peripherals/PeripheralBusAndroid.cpp
